# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Paper Burner X API
  version: 0.1.0
  description: Minimal spec skeleton. Extend per CODE_REVIEW plan.
servers:
  - url: /api
paths:
...
  /documents:
    get:
      summary: List documents
      parameters:
        - in: query
          name: page
          schema: { type: integer, minimum: 1 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100 }
      responses:
        '200': { description: OK }
    post:
      summary: Create document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                fileName: { type: string }
                fileSize: { type: integer }
                fileType: { type: string }
                status: { type: string }
      responses:
        '201': { description: Created }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []